+++ docs/changelog.rst
@@ -1,6 +1,6 @@
 Changelog
 =====================
-0.9.5 - la rosa de foc
+0.10 - la rosa de foc
 ----------------------
 
 Features
@@ -28,6 +28,7 @@ Bugfixes
 - `#8783 <https://0xacab.org/leap/bitmask-dev/issues/8783>`_: use username instead of provider in the vpn calls
 - `#8868 <https://0xacab.org/leap/bitmask-dev/issues/8868>`_: can't upload generated key with bitmask
 - `#8832 <https://0xacab.org/leap/bitmask-dev/issues/8832>`_: don't allow putting non-private keys for the keyring address
+- `#8901 <https://0xacab.org/leap/bitmask-dev/issues/8901>`_: use gpg1 binary if present
 
 Misc
 ~~~~